Move SDK callback definitions to a separate callback module
Followup to #54796. Since callbacks are going to be independent from
deadlines, we might as well move them to their own module.

I've essentially copied the code between files with the only actual
change being imports and switching to structlog.
