+++ airflow-core/tests/unit/models/test_deadline.py
@@ -28,7 +28,8 @@ from airflow.api_fastapi.core_api.datamodels.dag_run import DAGRunResponse
 from airflow.models import DagRun
 from airflow.models.deadline import Deadline, ReferenceModels, _fetch_from_db
 from airflow.providers.standard.operators.empty import EmptyOperator
-from airflow.sdk.definitions.deadline import AsyncCallback, DeadlineReference, SyncCallback
+from airflow.sdk.definitions.callback import AsyncCallback, SyncCallback
+from airflow.sdk.definitions.deadline import DeadlineReference
 from airflow.utils.state import DagRunState
 
 from tests_common.test_utils import db
